Improve input

+++ PlantPAN2.py
@@ -1,8 +1,37 @@
-# Author: Nattawet Sriwichai
-# Author-email: nattawet.sri@mail.kmutt.ac.th
-# https://github.com/evolu-tion/GenomeManagement
-# Copyright (c) 2016 Nattawet Sriwichai
-# License: LICENSE.txt 
+#!/usr/bin/env python
+""" 
+Copyright (c) 2016 King Mongkut's University technology Thonburi
+Author: Nattawet Sriwichai
+Contact: nattawet.sri@mail.kmutt.ac.th
+License: MIT License
+
+The MIT License
+
+Copyright (c) 2016 King Mongkut's University technology Thonburi
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in
+all copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+THE SOFTWARE. 
+
+
+Please cite the following paper:
+Chi-Nga Chow, Han-Qin Zheng, Nai-Yun Wu, Chia-Hung Chien, Hsien-Da Huang, Tzong-Yi Lee, Yi-Fan Chiang-Hsieh, Ping-Fu Hou, Tien-Yi Yang, and Wen-Chi Chang. 
+PlantPAN 2.0: an update of plant promoter analysis navigator for reconstructing transcriptional regulatory networks in plants, Nucleic Acids Res. 2015 : gkv1035v1-gkv1035.
+"""
 
 ############################### Initial Configuration ###########################################
 
@@ -23,7 +52,7 @@ filltered_strand = 'Yes'
 backup_file_path = "out/backup_promoter_analysis_PlantPAN/"
 file_name_tfbs_on_promoter = 'out/tfbs_on_promoter.txt'
 file_name_tfbs_on_promoter_CpG_island = 'out/tfbs_on_promoter_CpG_island.txt'
-file_name_tfbs_on_promoter_filltered = 'out/tfbs_on_promoter_filltered_[sim='+str(filltered_similar_cutoff)+'_strand_'+filltered_strand+'].txt'
+file_name_tfbs_on_promoter_filltered = 'out/tfbs_on_promoter_filltered.txt'
 
 ##################################################################################################
 
@@ -35,6 +64,7 @@ file_name_tfbs_on_promoter_filltered = 'out/tfbs_on_promoter_filltered_[sim='+st
 import urllib.parse
 import urllib.request
 import re
+import os
 
 def PlantPAN2(seq_name, sequence):
 	table_data = []
@@ -119,12 +149,20 @@ def PlantPAN2(seq_name, sequence):
 
 
 # Run PlantPAN2.0 of all sequence in fasta file
+if not os.path.exists(file_promoter):
+	print("Location of promoter sequence file is not correct")
+	exit()
+
+os.makedirs(os.path.dirname(backup_file_path), exist_ok=True)
+os.makedirs(os.path.dirname(file_name_tfbs_on_promoter), exist_ok=True)
 f = open(file_name_tfbs_on_promoter, 'a')
 f.write('Promoter_ID\tMatrix ID\tFamily\tPosition\tStrand\tSimilar Score\tHit Sequence\n')
 
+os.makedirs(os.path.dirname(file_name_tfbs_on_promoter_filltered), exist_ok=True)
 f_filltered = open(file_name_tfbs_on_promoter_filltered, 'w')
 f_filltered.write('Promoter_ID\tMatrix ID\tFamily\tPosition\tStrand\tSimilar Score\tHit Sequence\n')
 
+os.makedirs(os.path.dirname(file_name_tfbs_on_promoter_CpG_island), exist_ok=True)
 f_CpG_island = open(file_name_tfbs_on_promoter_CpG_island, 'w')
 f_CpG_island.write('Promoter_id\tBegin\tEnd\tLength\tGC freq\tCpG ratio\tAT skew\tCG skew\tStart-p\tStrand\tStrand-p\n')
