+++ promoter_retrieve.py
@@ -1,8 +1,32 @@
-# Author: Nattawet Sriwichai
-# Author-email: nattawet.sri@mail.kmutt.ac.th
-# https://github.com/evolu-tion/GenomeManagement
-# Copyright (c) 2016 Nattawet Sriwichai
-# License: LICENSE.txt 
+#!/usr/bin/env python
+""" 
+Copyright (c) 2016 King Mongkut's University technology Thonburi
+Author: Nattawet Sriwichai
+Contact: nattawet.sri@mail.kmutt.ac.th
+License: MIT License
+
+The MIT License
+
+Copyright (c) 2016 King Mongkut's University technology Thonburi
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in
+all copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+THE SOFTWARE. 
+"""
 
 ############################### Initial Configuration ###########################################
 # Genome input files including: 
@@ -24,7 +48,7 @@ start_promoter_from = 'TSS'
 upstream = 2000
 downstream = 0
 promoter_minimum_length = 500
-all_promoter_in_genome = 'All'
+all_promoter_in_genome = 'No'
 
 # Output file including:
 # 1) Promoter sequences
@@ -38,12 +62,25 @@ output_file_list_no_promoter = 'out/list_no_promoter.txt'
 
 ##################################### Python Main Program ########################################
 
-
+import os
+import sys
 from seq_manage import Fasta_manager
 from seq_manage import Gff_manager
 from seq_manage import Genome_manager
 
 def main():
+	if not os.path.exists(file_genome_seq):
+		print("Location of genome sequence file is not correct")
+		exit()
+	if not os.path.exists(file_gff):
+		print("Location of gene feature file is not correct")
+		exit()
+	if all_promoter_in_genome == 'No' and not os.path.exists(file_custom_list_of_gene):
+		print("Location of list of genes is not correct")
+		exit()
+
+	os.makedirs(os.path.dirname(output_file_promoter), exist_ok=True)
+	os.makedirs(os.path.dirname(output_file_list_no_promoter), exist_ok=True)
 	out_promoter = open(output_file_promoter, 'w')
 
 	if(all_promoter_in_genome == 'No'):
